refactor utils and controller

+++ src/main/java/com/example/oop6/models/field/commands/ResizeCommand.java
@@ -63,11 +63,11 @@ public class ResizeCommand implements Command{
         return "ResizeCommand";
     }
 
-    public void setDx(int dx) {
+    public void setDx(double dx) {
         this.dx = dx;
     }
 
-    public void setDy(int dy) {
+    public void setDy(double dy) {
         this.dy = dy;
     }
 }
